backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
djoser==2.1.0
python-dotenv==1.0.0
drf-yasg==1.21.5
